+++ transcendence/base/admin.py
@@ -1,0 +1,1 @@
+
