+++ transcendence/base/apps.py
@@ -1,0 +1,1 @@
+
