+++ transcendence/base/models.py
@@ -1,0 +1,1 @@
+
