+++ transcendence/base/templates/base.html
