+++ transcendence/base/templates/tournaments.html
