+++ transcendence/transcendence/settings.py
@@ -37,6 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'base',
     'user',
     'rest_framework',
     'rest_framework.authtoken',
